+++ e2e/elements/notification.element.ts
@@ -1,8 +1,19 @@
-import { by, element } from 'protractor';
+import { by, element, browser } from 'protractor';
 import { BaseElement } from './base.element';
 
 export class Notification extends BaseElement {
     private pageName: string;
+    private clickSectionScript = `
+    (function mouseEnter(el, etype) {
+    if (el.fireEvent) {
+        el.fireEvent('on' + etype);
+    } else {
+        var evObj = document.createEvent('Events');
+        evObj.initEvent(etype, true, false);
+        el.dispatchEvent(evObj);
+    }
+    })(arguments[0], 'mouseenter')
+    `;
 
     constructor(pageName: string) {
         super(by.tagName('simple-notification'));
@@ -51,7 +62,14 @@ export class Notification extends BaseElement {
         return this.assert('warn', message, header);
     }
 
+    private mouseOver() {
+        return browser.executeScript(this.clickSectionScript, this.getContainer());
+    }
+
     private async assert(type: string, message?: string, header?: string) {
+        if (await this.isVisible()) {
+            this.mouseOver();
+        }
         await expect(type === 'error' ? this.isError() : type === 'success' ? this.isSuccess() : type === 'warn' ? this.isWarning() : false)
             .toBe(true, `${this.pageName}: No Success notification message!`);
         if (message) {