+++ e2e/pages/administration/projectSettings.po/constants.ts
@@ -10,7 +10,8 @@ export const elements = {
     stepsSwitcher: new UiSwitch(by.css('#steps-feature ui-switch')),
     importCompareResultsPattern: new Input(by.css('#compare-result-pattern-feature input')),
     projectSelector: new Autocomplete(by.id('project-selector')),
-    saveFeatures: element(by.id('save-features'))
+    saveFeatures: element(by.id('save-features')),
+    stabilityResultsCount: new Input(by.css('#stability_count input'))
 };
 
 export const names = {