+++ e2e/pages/project/view.po/index.ts
@@ -15,7 +15,7 @@ class ProjectView extends BasePage {
         return elements.projectTitleLabel.getText();
     }
 
-    async isEmpty(){
+    async isEmpty() {
         return elements.emptyProjectMessage.isDisplayed();
     }
 }