+++ src/app/shared/models/test.ts
@@ -13,11 +13,7 @@ export class Test {
   developer_id?: number;
   results?: TestResult[];
   suites?: TestSuite[];
-  lastResultColors?: LastResulColors;
-}
-
-export class LastResulColors {
-  resolution_colors: string;
-  result_colors: string;
-  result_ids: string;
+  resolution_colors?: string;
+  result_colors?: string;
+  result_ids?: string;
 }